feat: admin view

+++ routes/web.php
@@ -3,7 +3,7 @@
 use Illuminate\Support\Facades\Route;
 
 // Home Route
-Route::redirect('/', '/products')->name('home.index');
+Route::get('/', 'App\Http\Controllers\HomeController@index')->name('home.index');
 
 // Review Routes
 Route::get('/reviews', 'App\Http\Controllers\ReviewController@index')->name('review.index');
@@ -21,21 +21,21 @@ Route::post('/orders/save', 'App\Http\Controllers\OrderController@save')->name('
 Route::get('/orders/{id}', 'App\Http\Controllers\OrderController@show')->name('order.show');
 Route::get('/orders/delete/{id}', 'App\Http\Controllers\OrderController@delete')->name('order.delete');
 
-// Product Routes
+// User Product Routes
 Route::get('/products', 'App\Http\Controllers\ProductController@index')->name('product.index');
-Route::get('/products/search', 'App\Http\Controllers\ProductController@search')->name('product.search');
-Route::get('/products/create', 'App\Http\Controllers\ProductController@create')->name('product.create');
-Route::post('/products/save', 'App\Http\Controllers\ProductController@save')->name('product.save');
-Route::get('/products/{id}/edit', 'App\Http\Controllers\ProductController@edit')->name('product.edit');
-Route::put('/products/{id}', 'App\Http\Controllers\ProductController@update')->name('product.update');
 Route::get('/products/{id}', 'App\Http\Controllers\ProductController@show')->name('product.show');
-Route::delete('/products/{id}', 'App\Http\Controllers\ProductController@destroy')->name('product.destroy');
 
-// Category Routes
-Route::get('/categories', 'App\Http\Controllers\CategoryController@index')->name('category.index');
-Route::get('/categories/create', 'App\Http\Controllers\CategoryController@create')->name('category.create');
-Route::post('/categories/save', 'App\Http\Controllers\CategoryController@save')->name('category.save');
-Route::get('/categories/{id}/edit', 'App\Http\Controllers\CategoryController@edit')->name('category.edit');
-Route::put('/categories/{id}', 'App\Http\Controllers\CategoryController@update')->name('category.update');
-Route::get('/categories/{id}', 'App\Http\Controllers\CategoryController@show')->name('category.show');
-Route::delete('/categories/{id}', 'App\Http\Controllers\CategoryController@destroy')->name('category.destroy');
+// Admin Routes
+Route::get('/admin/products', 'App\Http\Controllers\Admin\ProductController@index')->name('admin.product.index');
+Route::get('/admin/products/create', 'App\Http\Controllers\Admin\ProductController@create')->name('admin.product.create');
+Route::post('/admin/products', 'App\Http\Controllers\Admin\ProductController@store')->name('admin.product.store');
+Route::get('/admin/products/{id}/edit', 'App\Http\Controllers\Admin\ProductController@edit')->name('admin.product.edit');
+Route::put('/admin/products/{id}', 'App\Http\Controllers\Admin\ProductController@update')->name('admin.product.update');
+Route::delete('/admin/products/{id}', 'App\Http\Controllers\Admin\ProductController@destroy')->name('admin.product.destroy');
+
+Route::get('/admin/categories', 'App\Http\Controllers\Admin\CategoryController@index')->name('admin.category.index');
+Route::get('/admin/categories/create', 'App\Http\Controllers\Admin\CategoryController@create')->name('admin.category.create');
+Route::post('/admin/categories', 'App\Http\Controllers\Admin\CategoryController@store')->name('admin.category.store');
+Route::get('/admin/categories/{id}/edit', 'App\Http\Controllers\Admin\CategoryController@edit')->name('admin.category.edit');
+Route::put('/admin/categories/{id}', 'App\Http\Controllers\Admin\CategoryController@update')->name('admin.category.update');
+Route::delete('/admin/categories/{id}', 'App\Http\Controllers\Admin\CategoryController@destroy')->name('admin.category.destroy');
